Setup app data volume with /app/data of container
It turns out that mounting a host path in /app/data will shadow
the contents of the container /app/data. I expected docker to
copy all files from /app/data of container into the volume.
Since docker doesn't do this for us, we have to copy over the
contents ourselves before we start the container.

https://github.com/dotcloud/docker/issues/1992

Without this change, every app that has any 'initial' config
will have to basically write a setup script which has to run
on startup.

This changes is a little untested because volume mounting cannot
be tested on the Mac through boot2docker.

Notes:
1. The VOLUME cmd in Dockerfile merely exposes the directory to
be available for other container to mount as volumes (i.e they
will be available for --volumes-from)