Add a note that appdata delete doesn't work
docker has no concept of user namespace yet. This means that
whatever files are created by the container is seen on the host
appdata as a file by that user. If root is running in the container,
then the volume has files owned by root. This means that our app
code cannot delete the directoy anymore.

Note that some versions of docker even change the permissions of
the app's volume on host to root.root. This behavior apparently changes
depending on whether the volume is empty (!)

I tried playing around with setfacl/getfacl but that doesn't help
because the container can remove any ACL we set.

I see two options:
  - a setuid script
  - a sudo NOPASSWD script

Links to various issues:
https://github.com/dotcloud/docker/issues/2969
https://github.com/dotcloud/docker/issues/2360
https://github.com/dotcloud/docker/issues/2372
https://github.com/dotcloud/docker/issues/3124
https://github.com/dotcloud/docker/issues/2975
https://github.com/dotcloud/docker/issues/5189
https://github.com/dotcloud/docker/issues/1737