Use hyphen based subdomains for apps
Instead of
    appname.username.cloudron.us
we now use
    appname-username.cloudron.us

This allows us to get a wildcard certificate for cloudron.us and the
ssl stuff will work.

The motivation for appname-username instead of username-appname is primarily
UI at this point. We can show a suffix easily on the right side of a line edit.

Fixes #49