Add oauth scopes for each scoped route
All tokens are still given out with the
wildcard scope. The syncer and file related
routes don't have scopes. As well as the
firsttime use and provisioning routes.

We now have the following scopes so far:

- root      /api/v1/..
- profile   /api/v1/profile
- users     /api/v1/users/..
- apps      /api/v1/apps/..
- settings  /api/v1/settings/..

Each token can carry a comma separated list
of scopes, the token is good for.