Setup mysql addon
Also, move the deletion of containers into the installer script.
This is the place where supervisor is stopped, so it's a good idea
to stop the apps and containers there as well.

The actual container removal happens in postinstall because the
postinstall can decide to reuse apps if it wants