Allow specify regions for cloudron images
This reuses the 'regions' argument to transfer the
new image to the list specified. The first entry is the
region, where the temporary droplet is being created.

eg. --regions="ams3 ny2 sfo1"

The image droplet is created in Amsterdam.

The script currently does not wait for all images to be
fully transferred, yet.