Store app version in db at install time
It used to be saved in db at manifest download time. This is inherently
racy because the app version the user intended to download might be
different from what he ends up installing.

Same race applies to the update case.

This also ensures that we restore to the correct version during the restore
phase should the manifest never have gotten downloaded.

Part of #169