Make developerMode survive updates
Making developer mode survive upgrades can be done by storing
it in the db (and thus backed up). But this also makes it a
minor security issue - the user restores to a random backup
and he doesn't realize it's in developer mode.

It's best to keep this flag in cloudron.conf for now.

Fixes #265