make backupbox uploads snapshot
btrfs snapshotting does not really give us much in terms of atomicity
since the addons are dumped non-atomically.

The only way to make it all atomic is to stop the app. Snapshot and
let the app start asap. Because we don't stop apps when taking a backup (atm),
the whole btrfs snapshotting thing is superfluous.

part of #280