Change restart policy to always
It looks like containers with a restart policy of 'on-failure'
do not get restarted if docker crashes (guessing that it's because
docker doesn't record it as a container failure).

apps have a restart policy of 'on-failure' and they don't get restarted
whereas addons with a restart policy of 'always' do get restarted.

So, change restart policy of app to 'always' for now.