Do not backup apps that are not installed and running
Rest of the apps are in some intermediate state and we should
consider the data to be in some transient state that is not
worth backing up.

Note that we check the health and not the run state because the
app that is running could still be in middle of running some
db upgrade script. Healthy implies the app is up and running.