Pass old app config to app task
Girish 1 OCD 1

Creating a pure task based approach is not very easy with our current
approach. We use the app table entry to track status and all validation.
So, the approach of keeping old config in the table and passing around
a newConfig package is not happenning without touching lots of code.

The current approach simply stashes the old config in the apps table
itself. We could potentially move this to another 'tasks' table but
this then raises cleanup issues (should the taskmanager crash after
the apps has been removed...)

Fixes #416