Use cloudron.conf file to determine if this is an update
The installer determines if it an update based on existence of box dir.
It then calls the nginx splash setup code. The splash setup code relies
on the data directory being setup. Otherwise, it barfs in data/nginx.

This results in a case where restarting cloudron-installer when it is
unpacking the box code results in an error about being unable to write
admin.conf (since the data dir gets setup only in box setup.sh).

So, use cloudron.conf to determine if box was setup and running at some
point (this is already used in the js code)